refactored states into custom hooks
